[GSAM-167] Move operation definition and executor to internal/core/operations pkg
