+++ internal/core/operations/executor.go
@@ -1,6 +1,10 @@
-package operation
+package operations
 
-import "context"
+import (
+	"context"
+
+	"github.com/topfreegames/maestro/internal/core/entities/operation"
+)
 
 // Executor is where the actual operation logic is implemented, and it will
 // receive as input its correlated definition.
@@ -8,7 +12,7 @@ type Executor interface {
 	// This is where the operation logic will live; it will receive a context
 	// that will be used for deadline and cancelation. This function has only
 	// one return which is the operation error (if any);
-	Execute(ctx context.Context, operation *Operation, definition Definition) error
+	Execute(ctx context.Context, operation *operation.Operation, definition Definition) error
 
 	// This function is called if Execute returns an error. This will be used
 	// for operations that need to do some cleanup or any process if it fails.
