+++ internal/core/services/operation_manager/operation_manager.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 
 	"github.com/google/uuid"
+	"github.com/topfreegames/maestro/internal/core/operations"
 	"github.com/topfreegames/maestro/internal/core/entities/operation"
 	"github.com/topfreegames/maestro/internal/core/ports"
 	"github.com/topfreegames/maestro/internal/core/services/operations_registry"
@@ -29,7 +30,7 @@ func New(storage ports.OperationStorage) *OperationManager {
 	}
 }
 
-func (o *OperationManager) CreateOperation(ctx context.Context, definition operation.Definition) (*operation.Operation, error) {
+func (o *OperationManager) CreateOperation(ctx context.Context, definition operations.Definition) (*operation.Operation, error) {
 	op := &operation.Operation{
 		ID:             uuid.NewString(),
 		Status:         operation.StatusPending,
@@ -44,7 +45,7 @@ func (o *OperationManager) CreateOperation(ctx context.Context, definition opera
 	return op, nil
 }
 
-func (o *OperationManager) GetOperation(ctx context.Context, schedulerName, operationID string) (*operation.Operation, operation.Definition, error) {
+func (o *OperationManager) GetOperation(ctx context.Context, schedulerName, operationID string) (*operation.Operation, operations.Definition, error) {
 	op, definitionContents, err := o.storage.GetOperation(ctx, schedulerName, operationID)
 	if err != nil {
 		return nil, nil, err
@@ -64,7 +65,7 @@ func (o *OperationManager) GetOperation(ctx context.Context, schedulerName, oper
 }
 
 // NextSchedulerOperation returns the next scheduler operation to be processed.
-func (o *OperationManager) NextSchedulerOperation(ctx context.Context, schedulerName string) (*operation.Operation, operation.Definition, error) {
+func (o *OperationManager) NextSchedulerOperation(ctx context.Context, schedulerName string) (*operation.Operation, operations.Definition, error) {
 	operationID, err := o.storage.NextSchedulerOperationID(ctx, schedulerName)
 	if err != nil {
 		return nil, nil, fmt.Errorf("failed to retrieve the next operation: %w", err)
